+++ src/sll/api/util_api.c
@@ -1,7 +1,13 @@
 #include <sll/_sll_internal.h>
 #include <sll/api.h>
-#include <sll/static_object.h>
+#include <sll/assembly.h>
 #include <sll/types.h>
+
+
+
+__API_FUNC(util_instruction_count){
+	return sll_current_instruction_count;
+}